requirements.txt
streamlit
langchain
faiss-cpu 
langchain-core
langchain-google-genai 
PyPDF2
google-ai-generativelanguage
langchain-community
pandas
langchain-huggingface
faiss-cpu
wikipedia
sentence-transformers
transformers